+++ src/main/java/net/yxiao233/ifeu/api/networking/BooleanValueSyncS2C.java
@@ -1,6 +1,8 @@
 package net.yxiao233.ifeu.api.networking;
 
+import java.util.List;
+
 public interface BooleanValueSyncS2C {
-    void setValue(boolean value);
-    boolean getValues();
+    void setValue(List<Boolean> values);
+    List<Boolean> getValues();
 }