+++ src/main/java/net/yxiao233/ifeu/api/structure/IMultiBlockStructure.java
@@ -1,0 +1,5 @@
+package net.yxiao233.ifeu.api.structure;
+
+public interface IMultiBlockStructure {
+    MultiBlockStructure multiBlockStructure();
+}