+++ src/main/java/net/yxiao233/ifeu/common/block/entity/WeatherControllerEntity.java
@@ -146,7 +146,7 @@ public class WeatherControllerEntity extends IndustrialProcessingTile<WeatherCon
     public void serverTick(Level level, BlockPos pos, BlockState state, WeatherControllerEntity blockEntity) {
         super.serverTick(level, pos, state, blockEntity);
         if(!level.isClientSide()){
-            PacketDistributor.sendToAllPlayers((new BooleanSyncS2CPacket(getBlockPos(),finish)));
+            PacketDistributor.sendToAllPlayers((new BooleanSyncS2CPacket(getBlockPos(),List.of(finish))));
         }
     }
 
@@ -207,12 +207,12 @@ public class WeatherControllerEntity extends IndustrialProcessingTile<WeatherCon
 
 
     @Override
-    public void setValue(boolean value) {
-        this.finish = value;
+    public void setValue(List<Boolean> values) {
+        this.finish = values.getFirst();
     }
 
     @Override
-    public boolean getValues() {
-        return finish;
+    public List<Boolean> getValues() {
+        return List.of(finish);
     }
 }