+++ src/main/java/net/yxiao233/ifeu/common/registry/ModTags.java
@@ -14,6 +14,11 @@ public class ModTags {
         public static final TagKey<Block> NEEDS_DRAGON_STAR_TOOL = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:need_dragon_star_tool"));
         public static final TagKey<Block> MACHINE_FRAME_ULTIMATE = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:machine_frame/ultimate"));
         public static final TagKey<Block> WRENCH_PICKUP = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:wrench_pickup"));
+        public static final TagKey<Block> BLACK_HOLE_TANK = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:black_hole_tank"));
+        public static final TagKey<Block> BLACK_HOLE_CAPACITOR = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:black_hole_capacitor"));
+        public static final TagKey<Block> STORAGE_ITEM = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:storage/item"));
+        public static final TagKey<Block> STORAGE_FLUID = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:storage/fluid"));
+        public static final TagKey<Block> STORAGE_ENERGY = TagUtil.getBlockTag(ResourceLocation.parse("ifeu:storage/energy"));
         private static TagKey<Block> createTag(String name){
             return BlockTags.create(ResourceLocation.fromNamespaceAndPath(IndustrialForegoingExtraUpgrades.MODID,name));
         }